Generalized go.gd; added features to Spritesheet
go.gd has been generalized with a function or two that better handles more modules at once, instead of a bunch of functions each specialized to one.

spritesheet module now has play, stop, and pause for users, as well as a on_animation_end callback

The ancestor auto callback feature has been tweaked to not look for the closest ancestor with a script, but rather the closest ancestor whose script possesses the callback function.