RigidBody2D nodes now have go callbacks, and Area2D has a new callback for body collision
go_body_collision works on Area2D and RigidBody2D node ancestors, which runs when area_entered or body_entered occurs. In addition, if the RigidBody2D node ancestor has the callback function, it will enable contact monitor and contacts reporting, so the developer doesn't have to do this manually.