All nodes now connect all their signals to the nearest ancestor if the ancestor has the callback
Removed deprecated code
Deprecated go_collision and body_collision in favor of using the nodes' actual signal names, to keep users of GoGodot more in line with the actual Godot docs.
This update also allows every built-in signal to be automatically connected to an ancestor if the ancestor's script possesses a function that shares the name of the signal.